tornadofs/requirement.txt
absl-py                0.9.0       
astunparse             1.6.3       
asyncio                3.4.3       
cachetools             4.1.0       
certifi                2020.4.5.2  
chardet                3.0.4       
gast                   0.3.3       
google-auth            1.17.2      
google-auth-oauthlib   0.4.1       
google-pasta           0.2.0       
grpcio                 1.29.0      
h5py                   2.10.0      
idna                   2.9         
importlib-metadata     1.6.1       
Jinja2                 2.10.3      
joblib                 0.14.1      
kafka-python           2.0.0       
Keras                  2.4.1       
Keras-Preprocessing    1.1.2       
lxml                   4.4.2       
Markdown               3.2.2       
MarkupSafe             1.1.1       
numpy                  1.18.5      
oauthlib               3.1.0       
opt-einsum             3.2.1       
pandas                 0.24.2      
Pillow                 7.0.0       
pip                    19.3.1      
protobuf               3.12.2      
psycopg2               2.8.4       
py                     1.8.1       
pyasn1                 0.4.8       
pyasn1-modules         0.2.8       
PyMySQL                0.9.3       
pyshark                0.4.2.9     
python-dateutil        2.8.1       
pytz                   2020.1      
PyYAML                 5.3.1       
requests               2.24.0      
requests-oauthlib      1.3.0       
rsa                    4.6         
scikit-learn           0.22.2.post1
scipy                  1.4.1       
setproctitle           1.1.10      
setuptools             20.10.1     
six                    1.15.0      
sklearn                0.0         
SQLAlchemy             1.3.12      
tensorboard            2.2.2       
tensorboard-plugin-wit 1.6.0.post3 
tensorflow             2.2.0       
tensorflow-estimator   2.2.0       
termcolor              1.1.0       
tornado                6.0.3       
urllib3                1.25.9      
Werkzeug               1.0.1       
wheel                  0.34.2      
wrapt                  1.12.1      
zipp                   1.2.0 
